requirements_django.txt
Django==4.2
psycopg2-binary==2.9.6
django-extensions==3.2.3

# Optionally, add other libraries:
# tailwindcss, whitenoise, or anything else you use in your project
Pillow
djangorestframework
django-allauth
requests
PyJWT
cryptography